[Commands] Cleanup #emptyinventory Command. - Cleanup messages and logic. - Breakout #emptyinventory into its own command file.
